+++ src/tespy/networks/network_reader.py
@@ -41,7 +41,10 @@ from tespy.components import SubsystemInterface
 from tespy.components import Turbine
 from tespy.components import Valve
 from tespy.components import WaterElectrolyzer
+
 from tespy.components.newcomponents import *
+from tespy.components.AirDrier import *
+
 from tespy.connections import Bus
 from tespy.connections import Connection
 from tespy.connections import Ref
